Style : 예외처리 추가

+++ mung-BE/src/main/java/com/mung/mung/api/service/GameRoomService.java
@@ -1,6 +1,9 @@
 package com.mung.mung.api.service;
 
 import com.mung.mung.api.request.GameRoomCreateReq;
+import com.mung.mung.common.exception.custom.PlayerNotExistException;
+import com.mung.mung.common.exception.custom.RoomAlreadyExistsException;
+import com.mung.mung.common.exception.custom.RoomNotExistException;
 import com.mung.mung.db.entity.Game;
 import com.mung.mung.db.entity.GameRoom;
 import com.mung.mung.db.entity.Player;
@@ -22,16 +25,15 @@ public class GameRoomService {
     private final PlayerRepository playerRepository;
 
     @Transactional
-    public boolean makeRoom(String roomId, final GameRoomCreateReq gameRoomCreateReq){
+    public void makeRoom(String roomId, final GameRoomCreateReq gameRoomCreateReq){
         // DB에서 roomId가 있는지 검색 후 없으면 아래처럼 데이터 저장 후 생성, 있으면 return false
         if (roomId == null || roomId.isEmpty()) {
-
-            return false;
+            throw new RoomNotExistException();
         }
         GameRoom existingGameRoom = gameRoomRepository.findByRoomId(roomId);
         if (existingGameRoom != null) {
             // 이미 해당 roomId가 존재하는 경우
-            return false;
+            throw new RoomAlreadyExistsException();
         } else
         { // 존재하지 않으면 roomId로 방 생성
             ZoneId seoulZoneId = ZoneId.of("Asia/Seoul");
@@ -45,17 +47,15 @@ public class GameRoomService {
                     .startTime(seoulTime.toLocalDateTime())
                     .build();
             gameRoomRepository.save(gameRoom);
-            return true;
         }
     }
 
     @Transactional
     public boolean isPlayerExists(long playerId){
-        // DB에서 roomId가 있는지 검색 후 없으면 아래처럼 데이터 저장 후 생성, 있으면 return false
+        // DB에서 Player가 있는지 검색 후 없으면 아래처럼 데이터 저장 후 생성, 있으면 return false
         Player existingPlayer = playerRepository.findByPlayerId(playerId);
         if (existingPlayer == null) {
-            // 플레이어가 없으면
-            return false;
+            throw new PlayerNotExistException();
         } else
         {   // 있으면
             return true;
@@ -68,7 +68,7 @@ public class GameRoomService {
         GameRoom existingRoom = gameRoomRepository.findByRoomId(roomId);
         if (existingRoom == null) {
             // 방이 없으면
-            return false;
+            throw new RoomNotExistException();
         } else
         {
             return true;
@@ -79,23 +79,36 @@ public class GameRoomService {
     public String getRoomStatus(String roomId){
         // roomStatus 반환
        GameRoom gameRoom = gameRoomRepository.findByRoomId(roomId);
+       if (gameRoom == null){
+           throw new RoomNotExistException();
+       }
        return gameRoom.getStatus();
     }
 
 
     @Transactional
-    public void leaveRoom(long playerId){
+    public void leaveRoom(long playerId) {
+        if (playerRepository.findByPlayerId(playerId) == null){
+            throw new PlayerNotExistException();
+        }
         playerRepository.deleteByPlayerId(playerId);
     }
 
+
     @Transactional
     public void deleteRoom(String roomId){
+        if (gameRoomRepository.findByRoomId(roomId) == null){
+            throw new RoomNotExistException();
+        }
         gameRoomRepository.deleteByRoomId(roomId);
     }
 
     @Transactional
     public void roomInitialize(String roomId){
         GameRoom gameRoom = gameRoomRepository.findByRoomId(roomId);
+        if (gameRoom==null){
+            throw new RoomNotExistException();
+        }
         gameRoom.updateStatus("waiting");
     }
 }
